Refactoring (faucet-refactoring)
