+++ evm_loader/faucet/faucet.conf
@@ -1,4 +1,4 @@
-# Neon Token Faucet Service Config
+# Neon Faucet Service (example configugation)
 
 [rpc]
 bind = "0.0.0.0"
